refactor: renamed stores from "State" suffixes to "Store"
