+++ client/src/routes/editor/hooks/useDragDrop.ts
@@ -1,5 +1,5 @@
 import { DragEndEvent, DragStartEvent, PointerSensor, useSensor, useSensors } from "@dnd-kit/core";
-import { useEditorStore } from "../../../stores/editorState";
+import { useEditorStore } from "../../../stores/editorStore.ts";
 import { createSnapModifier } from "@dnd-kit/modifiers";
 import { Position } from "../types";
 import { SIZE_FACTOR } from "../constants";
