+++ client/src/routes/editor/hooks/useRotation.ts
@@ -1,4 +1,4 @@
-import { useEditorStore } from "../../../stores/editorState";
+import { useEditorStore } from "@/stores/editorStore.ts";
 import { SIZE_FACTOR } from "../constants";
 
 export function useRotation() {
